terraform/gcp/functions/gcs-s3-sync/requirements.txt
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.16.4
boto3==1.34.0